feat: :sparkles: 1. 同步 decimal 默认配置，将 precision 设置改为默认20;   2. getDecimalPlaces 部分逻辑优化（todo 有用例没通过，继续弄）

+++ src/constants/index.ts
@@ -1,7 +1,7 @@
 import Decimal from 'decimal.js'
 
 export const defaultDecimalConfigs: Decimal.Config = {
-  precision: 16, // 计算精度，参考 decimal.js 文档，可根据需求灵活调整
+  precision: 20, // 计算精度，参考 decimal.js 文档，可根据需求灵活调整
   rounding: Decimal.ROUND_HALF_UP, // 使用标准四舍五入 5进位 4舍去
   toExpNeg: -7,
   toExpPos: 21,
